Deduplicate test code
